# api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Qeet Pay API
  description: >-
    Qeet Pay — India-first payments, billing and GST infrastructure (UPI / Cards / NACH). Programmatic `/v1` calls
    authenticate with a `qp_live_` / `qp_test_` API key via the `X-Api-Key` header; money is always in integer minor
    units (paise).
  contact:
    name: Qeet Group
    url: https://qeet.in
  license:
    name: Proprietary
    url: https://qeet.in
  version: v1
servers:
  - url: https://api.pay.qeet.in
    description: Production
  - url: https://api.pay.staging.qeet.in
    description: Staging
  - url: http://localhost:4201
    description: Local (docker compose)
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/marketplace/sellers/{sellerId}/activate:
    post:
      tags:
        - Marketplace
      operationId: activate
      parameters:
        - name: sellerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SellerView'
  /v1/mandates:
    get:
      tags:
        - Mandates
      operationId: list_8
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MandateView'
    post:
      tags:
        - Mandates
      operationId: create_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMandateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MandateView'
  /v1/mandates/{id}/revoke:
    post:
      tags:
        - Mandates
      operationId: revoke
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MandateView'
  /v1/mandates/{id}/pause:
    post:
      tags:
        - Mandates
      operationId: pause
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MandateView'
  /v1/mandates/{id}/debit:
    post:
      tags:
        - Mandates
      operationId: debit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /v1/mandates/{id}/activate:
    post:
      tags:
        - Mandates
      operationId: activate_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MandateView'
...
  /v1/ledger/entries:
    post:
      tags:
        - Ledger
      operationId: postEntry
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostEntryRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /v1/mandates/{id}:
    get:
      tags:
        - Mandates
      operationId: get_9
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MandateView'
  /v1/mandates/{id}/debits:
    get:
      tags:
        - Mandates
      operationId: debits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MandateDebitView'
...
  /v1/ledger/accounts/{id}/balance:
    get:
      tags:
        - Ledger
      operationId: balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceView'
...
components:
  schemas:
...
    LineDto:
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        providerPaymentId:
          type: string
        grossMinor:
          type: integer
          format: int64
        feeMinor:
          type: integer
          format: int64
        taxMinor:
          type: integer
          format: int64
...
    SellerView:
      type: object
      properties:
        id:
          type: string
        sellerRef:
          type: string
        name:
          type: string
        gstin:
          type: string
        pan:
          type: string
        commissionBps:
          type: integer
          format: int32
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateMandateRequest:
      type: object
      properties:
        customerRef:
          type: string
          minLength: 1
        type:
          type: string
          enum:
            - UPI_AUTOPAY
            - NACH
        limitMinor:
          type: integer
          format: int64
        currency:
          type: string
        frequency:
          type: string
          enum:
            - DAILY
            - WEEKLY
            - MONTHLY
            - YEARLY
            - AS_PRESENTED
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - customerRef
        - frequency
        - startDate
        - type
    MandateView:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        type:
          type: string
        limitMinor:
          type: integer
          format: int64
        currency:
          type: string
        frequency:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
        providerMandateId:
          type: string
    DebitRequest:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        description:
          type: string
    ActivateRequest:
      type: object
      properties:
        providerMandateId:
          type: string
...
    PostEntryRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        currency:
          type: string
          minLength: 1
        lines:
          type: array
          items:
            $ref: '#/components/schemas/LineDto'
          minItems: 1
      required:
        - currency
        - description
        - lines
...
    MandateDebitView:
      type: object
      properties:
        id:
          type: string
        mandateId:
          type: string
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        ledgerEntryId:
          type: string
    BalanceView:
      type: object
      properties:
        accountId:
          type: string
        balanceMinor:
          type: integer
          format: int64
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Programmatic API key (`qp_live_…` / `qp_test_…`) presented as the `X-Api-Key` header.
      name: X-Api-Key
      in: header
...